protocol wip

+++ components/adc/adc.h
@@ -5,7 +5,7 @@
 #define AMP_SWITCH  static_cast<gpio_num_t>(14)
 
 void adc_task(void*);
-int adc_get();
+float adc_get();
 void dac_send(float x);
 void adc_init();
